Changes to save the config that was used for the run into the run folder.
Note that if you re-run with
same experiment name
but change the config, the old config still stays in the folder
so it could be confusing to do the training that way.

The intent is to use the sam
config and do the eval later.

And for a new training run,
always change the experiment name.